fix: handle all error events
When reading directories, the readable stream is not closed after
the first error. This is valid behavior because errors can occur while
reading the directory. For example, while reading a directory, another
process can delete it.

0) Previously, we could only handle the first error. After the first error,
the process crashed due to an unhandled exception. Now we will absorb
all errors from the readable stream.

1) Previously, we did not propagate errors to the merged stream.